+++ supertokens_python/recipe/thirdparty/api/implementation.py
@@ -17,9 +17,6 @@ from typing import TYPE_CHECKING, Union
 from urllib.parse import urlencode
 
 from httpx import AsyncClient
-from supertokens_python.recipe.thirdparty.constants import DEV_OAUTH_AUTHORIZATION_URL, DEV_OAUTH_REDIRECT_URL
-from supertokens_python.recipe.thirdparty.utils import is_using_oauth_development_client_id, \
-    get_actual_client_id_from_development_client_id
 from supertokens_python.exceptions import raise_general_exception
 from supertokens_python.recipe.session.asyncio import create_new_session
 from supertokens_python.recipe.thirdparty.interfaces import APIInterface, SignInUpPostOkResponse, \
@@ -29,6 +26,25 @@ if TYPE_CHECKING:
     from supertokens_python.recipe.thirdparty.interfaces import APIOptions, SignInUpPostResponse, \
         AuthorisationUrlGetResponse
     from supertokens_python.recipe.thirdparty.provider import Provider
+
+
+DEV_OAUTH_CLIENT_IDS = [
+    '1060725074195-kmeum4crr01uirfl2op9kd5acmi9jutn.apps.googleusercontent.com',  # google client id
+    '467101b197249757c71f'  # github client id
+]
+DEV_KEY_IDENTIFIER = "4398792-"
+DEV_OAUTH_AUTHORIZATION_URL = 'https://supertokens.io/dev/oauth/redirect-to-provider'
+DEV_OAUTH_REDIRECT_URL = 'https://supertokens.io/dev/oauth/redirect-to-app'
+
+
+def is_using_oauth_development_client_id(client_id: str):
+    return client_id.startswith(DEV_KEY_IDENTIFIER) or client_id in DEV_OAUTH_CLIENT_IDS
+
+
+def get_actual_client_id_from_development_client_id(client_id: str):
+    if client_id.startswith(DEV_KEY_IDENTIFIER):
+        return client_id.split(DEV_KEY_IDENTIFIER, 1)[1]
+    return client_id
 
 
 class APIImplementation(APIInterface):